requirements.txt
blinker==1.9.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
Flask==3.1.1
Flask-MySQL==1.6.0
idna==3.10
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
numpy==2.3.0
packaging==25.0
pandas==2.3.0
pluggy==1.6.0
Pygments==2.19.1
PyMySQL==1.1.1
pytest==8.4.0
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.4
six==1.17.0
tzdata==2025.2
urllib3==2.4.0
Werkzeug==3.1.3
